refactor: Consolidate memory caching logic and remove obsolete memory service files
