--- src-tauri/src/services/memory/mod.rs
@@ -1,9 +1,0 @@
-pub mod cached;
-pub mod dmidecode;
-pub mod memory_info_linux;
-pub mod proc_meminfo;
-
-pub use cached::*;
-pub use dmidecode::*;
-pub use memory_info_linux::*;
-pub use proc_meminfo::*;
