+++ src-tauri/src/utils/mod.rs
@@ -1,6 +1,8 @@
 pub mod color;
 pub mod file;
 pub mod formatter;
-pub mod ip;
 pub mod logger;
 pub mod tauri;
+
+#[cfg(target_os = "windows")]
+pub mod ip;
